+++ zpr/src/vsapi_types/util/ip.rs
@@ -1,0 +1,16 @@
+use crate::vsapi_types::VsapiTypeError;
+use std::net::IpAddr;
+
+pub fn ip_addr_from_vec(v: Vec<u8>) -> Result<IpAddr, VsapiTypeError> {
+    match v.len() {
+        4 => Ok(IpAddr::from(
+            <[u8; 4]>::try_from(v.as_slice()).expect("Bad IP length"),
+        )),
+        16 => Ok(IpAddr::from(
+            <[u8; 16]>::try_from(v.as_slice()).expect("Bad IP length"),
+        )),
+        _ => Err(VsapiTypeError::DeserializationError(
+            "Bad IP Address format",
+        )),
+    }
+}
